requirements.txt
PyQt5
pythonocc-core